+++ Hatify/Hatify.cs
@@ -1,4 +1,3 @@
-using System.Collections;
 using static BodyModelAdditionsAPI.Main;
 using BepInEx;
 using BepInEx.Configuration;
@@ -11,7 +10,7 @@ using System.Collections.Generic;
 
 namespace Hatify
 {
-  [BepInPlugin("com.Nuxlar.Hatify", "Hatify", "1.2.0")]
+  [BepInPlugin("com.Nuxlar.Hatify", "Hatify", "1.3.0")]
 
   public class Hatify : BaseUnityPlugin
   {
@@ -31,7 +30,12 @@ namespace Hatify
       "EngiBody",
       "EngiTurretBody",
       "EngiWalkerTurretBody",
-      "LoaderBody"
+      "LoaderBody",
+      "SeekerBody",
+      "FalseSonBody",
+      "ChefBody",
+      "TreebotBody",
+      "ToolbotBody"
     };
 
     public static ConfigEntry<float> commandoSize;
@@ -47,6 +51,11 @@ namespace Hatify
     public static ConfigEntry<float> captainSize;
     public static ConfigEntry<float> railgunnerSize;
     public static ConfigEntry<float> fiendSize;
+    public static ConfigEntry<float> seekerSize;
+    public static ConfigEntry<float> rexSize;
+    public static ConfigEntry<float> chefSize;
+    public static ConfigEntry<float> falseSonSize;
+    public static ConfigEntry<float> multSize;
     private static ConfigFile HatifyConfig { get; set; }
 
     public void Awake()
@@ -65,6 +74,11 @@ namespace Hatify
       captainSize = HatifyConfig.Bind<float>("General", "Captain Hat Size", 1.3f, "The scale of the hat.");
       railgunnerSize = HatifyConfig.Bind<float>("General", "Railgunner Hat Size", 1f, "The scale of the hat.");
       fiendSize = HatifyConfig.Bind<float>("General", "Fiend Hat Size", 1.4f, "The scale of the hat.");
+      seekerSize = HatifyConfig.Bind<float>("General", "Seeker Hat Size", 1f, "The scale of the hat.");
+      rexSize = HatifyConfig.Bind<float>("General", "REX Hat Size", 2f, "The scale of the hat.");
+      chefSize = HatifyConfig.Bind<float>("General", "CHEF Hat Size", 2f, "The scale of the hat.");
+      falseSonSize = HatifyConfig.Bind<float>("General", "False Son Hat Size", 2.5f, "The scale of the hat.");
+      multSize = HatifyConfig.Bind<float>("General", "MUL-T Hat Size", 15f, "The scale of the hat.");
 
       LoadAssets();
     }
@@ -82,13 +96,87 @@ namespace Hatify
       {
         GameObject actualHat = x.Result.transform.GetChild(4).GetChild(2).GetChild(0).GetChild(6).GetChild(0).GetChild(2).GetChild(0).gameObject;
         actualHat.AddComponent<NetworkIdentity>();
+        AssetReferenceT<SkinDef> skinDefRef = new AssetReferenceT<SkinDef>(RoR2BepInExPack.GameAssetPaths.RoR2_Base_Bandit2.skinBandit2AltColossus_asset);
+        SkinDef banditColossusSkin = AssetAsyncReferenceManager<SkinDef>.LoadAsset(skinDefRef).WaitForCompletion();
+        AssetReferenceT<SkinDef> skinDefRef2 = new AssetReferenceT<SkinDef>(RoR2BepInExPack.GameAssetPaths.RoR2_DLC2_Chef.skinChefAlt_asset);
+        SkinDef chefSkin = AssetAsyncReferenceManager<SkinDef>.LoadAsset(skinDefRef2).WaitForCompletion();
 
         foreach (string bodyName in bodyNames)
         {
           Vector3 hatSize = Vector3.zero;
-
           switch (bodyName)
           {
+            case "SeekerBody":
+              hatSize = new Vector3(seekerSize.Value, seekerSize.Value, seekerSize.Value);
+              if (hatSize != Vector3.zero)
+              {
+                ModelPartInfo modelPartInfo = new ModelPartInfo
+                {
+                  bodyName = bodyName,
+                  gameObject = actualHat,
+                  inputString = "Head",
+                  codeAfterApplying = PlaceItSeeker,
+                };
+                new ModelPart(modelPartInfo);
+              }
+              break;
+            case "FalseSonBody":
+              hatSize = new Vector3(falseSonSize.Value, falseSonSize.Value, falseSonSize.Value);
+              if (hatSize != Vector3.zero)
+              {
+                ModelPartInfo modelPartInfo = new ModelPartInfo
+                {
+                  bodyName = bodyName,
+                  gameObject = actualHat,
+                  inputString = "Head",
+                  codeAfterApplying = PlaceItFalseSon,
+                };
+                new ModelPart(modelPartInfo);
+              }
+              break;
+            case "ChefBody":
+              hatSize = new Vector3(chefSize.Value, chefSize.Value, chefSize.Value);
+              if (hatSize != Vector3.zero)
+              {
+                ModelPartInfo modelPartInfo = new ModelPartInfo
+                {
+                  bodyName = bodyName,
+                  gameObject = actualHat,
+                  inputString = "Head",
+                  codeAfterApplying = PlaceItChef,
+                  skinDef = chefSkin
+                };
+                new ModelPart(modelPartInfo);
+              }
+              break;
+            case "ToolbotBody":
+              hatSize = new Vector3(multSize.Value, multSize.Value, multSize.Value);
+              if (hatSize != Vector3.zero)
+              {
+                ModelPartInfo modelPartInfo = new ModelPartInfo
+                {
+                  bodyName = bodyName,
+                  gameObject = actualHat,
+                  inputString = "Head",
+                  codeAfterApplying = PlaceItMult,
+                };
+                new ModelPart(modelPartInfo);
+              }
+              break;
+            case "TreebotBody":
+              hatSize = new Vector3(rexSize.Value, rexSize.Value, rexSize.Value);
+              if (hatSize != Vector3.zero)
+              {
+                ModelPartInfo modelPartInfo = new ModelPartInfo
+                {
+                  bodyName = bodyName,
+                  gameObject = actualHat,
+                  inputString = "FlowerBase",
+                  codeAfterApplying = PlaceItRex,
+                };
+                new ModelPart(modelPartInfo);
+              }
+              break;
             case "Bandit2Body":
               hatSize = new Vector3(banditSize.Value, banditSize.Value, banditSize.Value);
               if (hatSize != Vector3.zero)
@@ -99,6 +187,7 @@ namespace Hatify
                   gameObject = actualHat,
                   inputString = "Head",
                   codeAfterApplying = PlaceItBandit,
+                  skinDef = banditColossusSkin
                 };
                 new ModelPart(modelPartInfo);
               }
@@ -277,7 +366,44 @@ namespace Hatify
         }
       };
     }
+    static void PlaceItMult(GameObject modelObject, ChildLocator childLocator, CharacterModel characterModel, ActivePartsComponent activePartsComponent)
+    {
+      modelObject.transform.localPosition = new Vector3(0f, 1.5f, 1.7f);
+      modelObject.transform.localScale = new Vector3(multSize.Value, multSize.Value, multSize.Value);
+      modelObject.transform.Rotate(new Vector3(315f, 180f, 0.0f));
+      modelObject.transform.GetChild(0).GetComponent<MeshRenderer>().material = hatMat;
+    }
+    static void PlaceItChef(GameObject modelObject, ChildLocator childLocator, CharacterModel characterModel, ActivePartsComponent activePartsComponent)
+    {
+      Transform chefHat = characterModel.transform.Find("meshChefPizzaCutter");
+      if (chefHat)
+        chefHat.gameObject.SetActive(false);
 
+      modelObject.transform.localPosition = new Vector3(-0.35f, 0f, 0f);
+      modelObject.transform.localScale = new Vector3(chefSize.Value, chefSize.Value, chefSize.Value);
+      modelObject.transform.Rotate(new Vector3(60f, 90f, 180f));
+      modelObject.transform.GetChild(0).GetComponent<MeshRenderer>().material = hatMat;
+    }
+    static void PlaceItFalseSon(GameObject modelObject, ChildLocator childLocator, CharacterModel characterModel, ActivePartsComponent activePartsComponent)
+    {
+      modelObject.transform.localPosition = new Vector3(0f, 0.34f, -0.05f);
+      modelObject.transform.localScale = new Vector3(falseSonSize.Value, falseSonSize.Value, falseSonSize.Value);
+      modelObject.transform.Rotate(new Vector3(10f, 0.0f, 0.0f));
+      modelObject.transform.GetChild(0).GetComponent<MeshRenderer>().material = hatMat;
+    }
+    static void PlaceItSeeker(GameObject modelObject, ChildLocator childLocator, CharacterModel characterModel, ActivePartsComponent activePartsComponent)
+    {
+      modelObject.transform.localPosition = new Vector3(0f, 0.2f, 0f);
+      modelObject.transform.localScale = new Vector3(seekerSize.Value, seekerSize.Value, seekerSize.Value);
+      modelObject.transform.GetChild(0).GetComponent<MeshRenderer>().material = banditHatMat;
+    }
+    static void PlaceItRex(GameObject modelObject, ChildLocator childLocator, CharacterModel characterModel, ActivePartsComponent activePartsComponent)
+    {
+      modelObject.transform.localPosition = new Vector3(0f, 2f, 0f);
+      modelObject.transform.localScale = new Vector3(rexSize.Value, rexSize.Value, rexSize.Value);
+      modelObject.transform.Rotate(new Vector3(10f, 0.0f, 0.0f));
+      modelObject.transform.GetChild(0).GetComponent<MeshRenderer>().material = hatMat;
+    }
     static void PlaceItBandit(GameObject modelObject, ChildLocator childLocator, CharacterModel characterModel, ActivePartsComponent activePartsComponent)
     {
       modelObject.transform.localPosition = new Vector3(0f, 0.15f, 0f);
@@ -287,6 +413,13 @@ namespace Hatify
     }
     static void PlaceItCaptain(GameObject modelObject, ChildLocator childLocator, CharacterModel characterModel, ActivePartsComponent activePartsComponent)
     {
+      Transform armature = characterModel.transform.Find("CaptainArmature");
+      if (armature)
+      {
+        Transform captainHat = armature.GetChild(0).GetChild(0).GetChild(2).GetChild(0).GetChild(2).GetChild(0).GetChild(4);
+        if (captainHat)
+          captainHat.gameObject.SetActive(false);
+      }
       modelObject.transform.localPosition = new Vector3(0f, 0.15f, 0f);
       modelObject.transform.localScale = new Vector3(captainSize.Value, captainSize.Value, captainSize.Value);
       modelObject.transform.Rotate(new Vector3(15f, 0.0f, 0.0f));
@@ -301,7 +434,7 @@ namespace Hatify
     }
     static void PlaceItRailgunner(GameObject modelObject, ChildLocator childLocator, CharacterModel characterModel, ActivePartsComponent activePartsComponent)
     {
-      modelObject.transform.localPosition = new Vector3(0.0f, 0.175f, -0.025f);
+      modelObject.transform.localPosition = new Vector3(0.0f, 0.18f, -0.025f);
       modelObject.transform.localScale = new Vector3(railgunnerSize.Value, railgunnerSize.Value, railgunnerSize.Value);
       modelObject.transform.Rotate(new Vector3(30f, 0.0f, 0.0f));
       modelObject.transform.GetChild(0).GetComponent<MeshRenderer>().material = hatMat;